Move some files around and run clippy
