+++ programs/drift/src/instructions/keeper.rs
@@ -10,6 +10,8 @@ use crate::load_mut;
 use crate::math::constants::QUOTE_SPOT_MARKET_INDEX;
 use crate::math::insurance::if_shares_to_vault_amount;
 use crate::math::spot_withdraw::validate_spot_market_vault_amount;
+use crate::state::fulfillment_params::drift::MatchFulfillmentParams;
+use crate::state::fulfillment_params::serum::SerumFulfillmentParams;
 use crate::state::insurance_fund_stake::InsuranceFundStake;
 use crate::state::oracle_map::OracleMap;
 use crate::state::perp_market::{MarketStatus, PerpMarket};
@@ -17,9 +19,7 @@ use crate::state::perp_market_map::{
     get_market_set_for_user_positions, get_market_set_from_list, get_writable_perp_market_set,
     MarketSet, PerpMarketMap,
 };
-use crate::state::spot_fulfillment_params::{
-    MatchFulfillmentParams, SerumFulfillmentParams, SpotFulfillmentParams,
-};
+use crate::state::spot_fulfillment_params::SpotFulfillmentParams;
 use crate::state::spot_market::SpotMarket;
 use crate::state::spot_market_map::{
     get_writable_spot_market_set, get_writable_spot_market_set_from_many,
@@ -1194,7 +1194,7 @@ pub fn handle_settle_revenue_to_insurance_fund(
         &ctx.accounts.insurance_fund_vault,
         &ctx.accounts.drift_signer,
         state.signer_nonce,
-        token_amount as u64,
+        token_amount,
     )?;
 
     // reload the spot market vault balance so it's up-to-date
