+++ programs/drift/src/math/constants.rs
@@ -82,7 +82,7 @@ pub const AMM_TO_QUOTE_PRECISION_RATIO_I128: i128 =
 pub const AMM_TIMES_PEG_TO_QUOTE_PRECISION_RATIO: u128 =
     AMM_RESERVE_PRECISION * PEG_PRECISION / QUOTE_PRECISION; // expo: 9
 pub const QUOTE_TO_BASE_AMT_FUNDING_PRECISION: i128 =
-    (AMM_RESERVE_PRECISION_I128 * FUNDING_RATE_PRECISION_I128 / QUOTE_PRECISION_I128) as i128; // expo: 12
+    AMM_RESERVE_PRECISION_I128 * FUNDING_RATE_PRECISION_I128 / QUOTE_PRECISION_I128; // expo: 12
 pub const PRICE_TO_QUOTE_PRECISION_RATIO: u128 = PRICE_PRECISION / QUOTE_PRECISION; // expo: 1
 pub const PRICE_TIMES_AMM_TO_QUOTE_PRECISION_RATIO: u128 =
     PRICE_PRECISION * AMM_TO_QUOTE_PRECISION_RATIO; // expo 9
@@ -137,9 +137,9 @@ pub const ONE_THOUSAND_QUOTE: u64 = 1_000_u64 * QUOTE_PRECISION_U64;
 pub const TWO_HUNDRED_FIFTY_QUOTE: u64 = 250_u64 * QUOTE_PRECISION_U64;
 
 // INSURANCE TIERS
-pub const INSURANCE_A_MAX: u64 = ONE_HUNDRED_MILLION_QUOTE as u64;
-pub const INSURANCE_B_MAX: u64 = ONE_HUNDRED_THOUSAND_QUOTE as u64;
-pub const INSURANCE_C_MAX: u64 = TWENTY_FIVE_THOUSAND_QUOTE as u64;
+pub const INSURANCE_A_MAX: u64 = ONE_HUNDRED_MILLION_QUOTE;
+pub const INSURANCE_B_MAX: u64 = ONE_HUNDRED_THOUSAND_QUOTE;
+pub const INSURANCE_C_MAX: u64 = TWENTY_FIVE_THOUSAND_QUOTE;
 pub const INSURANCE_SPECULATIVE_MAX: u64 = 0;
 
 // QUOTE THRESHOLDS
@@ -174,8 +174,8 @@ pub const MAX_LIQUIDATION_SLIPPAGE: i128 = 10_000; // expo = -2
 pub const MAX_LIQUIDATION_SLIPPAGE_U128: u128 = 10_000; // expo = -2
 pub const MAX_MARK_TWAP_DIVERGENCE: u128 = 500_000; // expo = -3
 
-pub const MAX_MARGIN_RATIO: u32 = MARGIN_PRECISION as u32; // 1x leverage
-pub const MIN_MARGIN_RATIO: u32 = MARGIN_PRECISION as u32 / 50; // 50x leverage
+pub const MAX_MARGIN_RATIO: u32 = MARGIN_PRECISION; // 1x leverage
+pub const MIN_MARGIN_RATIO: u32 = MARGIN_PRECISION / 50; // 50x leverage
 
 pub const MAX_BID_ASK_INVENTORY_SKEW_FACTOR: u64 = 10 * BID_ASK_SPREAD_PRECISION;
 
@@ -185,7 +185,7 @@ pub const DEFAULT_MAX_TWAP_UPDATE_PRICE_BAND_DENOMINATOR: i64 = 3; // '3' here m
 // DEFAULTS
 pub const DEFAULT_REVENUE_SINCE_LAST_FUNDING_SPREAD_RETREAT: i64 = -25 * QUOTE_PRECISION_I64; //$25 loss
 pub const DEFAULT_LARGE_BID_ASK_FACTOR: u64 = 10 * BID_ASK_SPREAD_PRECISION;
-pub const DEFAULT_LIQUIDATION_MARGIN_BUFFER_RATIO: u32 = (MARGIN_PRECISION as u32) / 50; // 2%
+pub const DEFAULT_LIQUIDATION_MARGIN_BUFFER_RATIO: u32 = MARGIN_PRECISION / 50; // 2%
 pub const DEFAULT_BASE_ASSET_AMOUNT_STEP_SIZE: u64 = BASE_PRECISION_U64 / 10000; // 1e-4;
 pub const DEFAULT_QUOTE_ASSET_AMOUNT_TICK_SIZE: u64 =
     PRICE_PRECISION_U64 / DEFAULT_BASE_ASSET_AMOUNT_STEP_SIZE; // 1e-2
